add cmd version
this means that you'll be able to use this without needing to explicitly
be inside a git bash shell (for example when you are using msysgit through
something like cmder).

because cmd doesn't support the notion of a bashrc to perform an
equivalent to source-ing the sh file like install.sh can do, we instead
put a bat into C:\Windows, which generally has a good chance of existing
and being on PATH. That just forwards arguments back into the install dir,
which works, although it's somewhat garish.